fix: 🐛 copy stephen emoji parser changes n fix emoji parser in slack emoji
